applications: nrf5340_audio: Add codec_helper
codec_helper created.
