+++ applications/nrf5340_audio/src/bluetooth/CMakeLists.txt
@@ -6,6 +6,7 @@
 
 target_sources(app PRIVATE
 	       ${CMAKE_CURRENT_SOURCE_DIR}/ble_core.c
+	       ${CMAKE_CURRENT_SOURCE_DIR}/codec_helper.c
 )
 
 if (CONFIG_TRANSPORT_CIS)
